Update to v2.34.0

+++ docker/scripts/utils.sh
@@ -19,9 +19,10 @@ function wait_file() {
     logd "waiting for $file..."
     sleep 5;
     count=$((count + 1))
-    [[ $count -ge 50 ]] \
-      && loge "Waited too long, aborting..." && exit 1 \
-      || :
+    if [[ "$count" -ge 50 ]]; then
+      loge "Waited too long, aborting..."
+      exit 1
+    fi
   done
 }
 
@@ -35,7 +36,7 @@ function wait_for_one() {
 }
 
 function exit_if_empty() {
-  [ -z "$1" ] && loge "$2" && exit 1 || :
+  if [ -z "$1" ]; then loge "$2"; exit 1; fi
 }
 
 function remove_sync_files() {
@@ -77,15 +78,19 @@ function maybe_install_app() {
 function drop_logs() {
   case $TEE in
     vanilla )
-      [[ ! $(rm -r "$FOLDER_PROXY/logs" 2> /dev/null) ]] \
-        && logi "Failed to drop the logs, maybe folder doesn't exist..." \
-        || logi "Dropping logs...done!"
+      if rm -r "$FOLDER_PROXY/logs" 2> /dev/null; then
+        logi "Dropping logs...done!"
+      else
+        logi "Failed to drop the logs, maybe folder doesn't exist..."
+      fi
       ;;
   esac 
 }
 
 function drop_sync_files() {
-  rm "$FOLDER_SYNC"/*.json 2> /dev/null
+  if [[ -z "$SKIP_INIT_BLOCK" ]]; then
+    rm "$FOLDER_SYNC"/*.json 2> /dev/null  
+  fi
   rm "$FOLDER_SYNC"/*.start 2> /dev/null
 
   logi "Removing sync files...done!"
@@ -130,56 +135,24 @@ function touch_start_files() {
 }
 
 function check_safetyfile_exists() {
-  [[ ! -f $FILE_SAFETY ]] \
-    && loge "Safety file not defined, aborting!" \
-    && exit 1 \
-    || :
+  if [[ ! -f $FILE_SAFETY ]]; then
+    loge "Safety file not defined, aborting!"
+    exit 1
+  fi
 }
 
 function is_native() {
-  local symbol
-
-  [[ "$NATIVE_SYMBOL" == "$symbol" ]] \
-    && return 0 \
-    || return 1
+  local regex
+  regex=$1
+  echo "$NATIVE_SYMBOL" | grep -E -o "$regex"
 }
 
 function is_host() {
-  local symbol
-
-  [[ "$HOST_SYMBOL" == "$symbol" ]] \
-    && return 0 \
-    || return 1
+  local regex
+  regex=$1
+  echo "$HOST_SYMBOL" | grep -E -o "$regex"
 }
 
-function add_bridge_info() {
-  local public_key
-  local api_server_file
-  local smart_contract_address
-  
-  api_server_file=$FOLDER_SYNC/api-server.json
-
-  # TODO: this supports pbtc-on-eth only
-
-  if [[ -f "$api_server_file" ]]; then
-    # native_address=$(jq -r '.ENCLAVE_PUBLIC_KEY' "$api_server_file")
-    smart_contract_address=$(jq -r '.SMART_CONTRACT_ADDRESS' "$api_server_file")
-
-    bridge='{
-      "p'$NATIVE_SYMBOL'-on-'$HOST_SYMBOL'": {
-        "NATIVE_TOKEN_ADDRESS": "'$public_key'",
-        "HOST_TOKEN_ADDRESS": "'$smart_contract_address'",
-        "NATIVE_SYMBOL": "'$NATIVE_SYMBOL'",
-        "HOST_SYMBOL": "'$HOST_SYMBOL'"
-      }
-    }'
-    
-    new_api_json=$(jq ".BRIDGES=$bridge" "$api_server_file")
-
-    logd "$new_api_json"
-    echo "$new_api_json" > "$api_server_file"
-  fi
-}
 
 function maybe_install_proxy_deps() {
   case $TEE in
@@ -209,4 +182,93 @@ function check_enclave_is_ready() {
       logi "Nitro enclave is listening!"
       ;;
   esac 
+}
+
+function get_symbol() {
+  local regexp
+  regexp=$1
+  __symbol=$2
+  
+  symbol=$(echo "$NATIVE_SYMBOL $HOST_SYMBOL" | grep -E -o "$regexp")
+
+  if [[ -z "$symbol" ]]; then
+    loge "Failed to get the symbol using this regexp: $regexp"
+    exit 1
+  fi
+
+  # shellcheck disable=SC2140
+  eval "$__symbol"="'$symbol'"
+}
+
+function get_type() {
+  local symbol
+  symbol=$1
+  __type=$2
+
+  type=null
+  if [[ "$NATIVE_SYMBOL" == "$symbol" ]]; then
+    type=native
+  elif [[ "$HOST_SYMBOL" == "$symbol" ]]; then
+    type=host
+  else
+    loge "Unable to find the type for $symbol"
+    exit 1
+  fi
+
+  # shellcheck disable=SC2140
+  eval "$__type"="'$type'"
+}
+
+function get_reduced_bridge_type() {
+  local bridge_type
+  bridge_type=$1
+
+  __reduced_bridge_type=$2
+
+  reduced_bridge_type=$(jq -r ".\"$bridge_type\"" <<< "$REDUCED_BRIDGE_TYPE_MAPPING")
+
+  # shellcheck disable=SC2140
+  eval "$__reduced_bridge_type"="'$reduced_bridge_type'"
+}
+
+function get_native_symbol_from_bridge_type() {
+  local bridge_type
+  bridge_type=$1
+  __native_symbol=$2
+
+  native_symbol=$(\
+    grep -E -o 'p[a-z0-9]{3,}' <<< "$bridge_type" \
+      | awk '{print substr($1, 2)}' \
+  )
+
+  # shellcheck disable=SC2140
+  eval "$__native_symbol"="'$native_symbol'"
+}
+
+function get_host_symbol_from_bridge_type() {
+  local bridge_type
+  bridge_type=$1
+  __host_symbol=$2
+
+  host_symbol=$(\
+    grep -E -o '\-[a-z0-9]{3,}' <<< "$bridge_type" \
+      | awk '{print substr($1, 2)}' \
+  )
+
+  # shellcheck disable=SC2140
+  eval "$__host_symbol"="'$host_symbol'"
+}
+
+function get_or_exit() {
+  configurable=$1
+  __conf_value=$2
+
+  # shellcheck disable=SC2140  
+  eval "value"="\$$configurable"
+
+  # shellcheck disable=SC2154  
+  exit_if_empty "$value" "Value for $configurable is missing!"
+
+  # shellcheck disable=SC2140
+  eval "$__conf_value"="'$value'"
 }
